+++ Assets/Scripts/Battlefield/CharacterFactory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fbcb87d1ac398d64ba3ded402741d2ad
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
